# docs/swagger.yaml
definitions:
...
  dtos.LoginUserInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  dtos.ProductCategoryFormatter:
    properties:
      category_id:
        type: integer
      description:
        type: string
      product_name:
        type: string
    type: object
  dtos.ProductDetailFormatter:
    properties:
      category:
        $ref: '#/definitions/dtos.ProductCategoryFormatter'
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      price:
        type: number
      product_name:
        type: string
      quantity:
        type: integer
      slug:
        type: string
      user:
        $ref: '#/definitions/dtos.ProductUserFormatter'
      user_id:
        type: integer
    type: object
  dtos.ProductFormatter:
    properties:
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      price:
        type: number
      product_name:
        type: string
      quantity:
        type: integer
      slug:
        type: string
      user_id:
        type: integer
    type: object
  dtos.ProductUserFormatter:
    properties:
      full_name:
        type: string
      role:
        type: string
      user_id:
        type: integer
    type: object
  dtos.RegisterUserInput:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
    type: object
  dtos.UserRegisterFormatter:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
  utils.Meta:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/utils.Meta'
    type: object
info:
  contact:
    email: nadapdaponai21@gmail.com
    name: API Support
    url: https://onainadapdap1.github.io/
  description: This is service for my Online Store API assignment.
  termsOfService: http://swagger.io/terms/
  title: Online Store API
  version: "1.0"
paths:
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: User login with email and password
      parameters:
      - description: Login user input
        in: body
        name: dtos.LoginUserInput
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.UserRegisterFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Login user
      tags:
      - users
...
  /api/v1/products:
    get:
      description: Get all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.ProductDetailFormatter'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all products
      tags:
      - products
  /api/v1/products/product:
    post:
      consumes:
      - multipart/form-data
      description: Create a new Product with a given name, description and image file
      parameters:
      - description: Name of the Product
        in: formData
        name: name
        required: true
        type: string
      - description: Description of the product
        in: formData
        name: description
        required: true
        type: string
      - description: price of the product
        in: formData
        name: price
        required: true
        type: number
      - description: quantity of the product
        in: formData
        name: quantity
        required: true
        type: integer
      - description: category id of the product
        in: formData
        name: category_id
        required: true
        type: integer
      - description: Image file of the product
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ProductFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create Product
      tags:
      - products
  /api/v1/products/product/{slug}:
    get:
      description: Get one product by slug
      parameters:
      - description: get product by slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ProductDetailFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get one product by slug
      tags:
      - products
    put:
      consumes:
      - application/json
      - multipart/form-data
      description: Update product
      parameters:
      - description: update category by slug
        in: path
        name: slug
        required: true
        type: string
      - description: Name of the Product
        in: formData
        name: name
        required: true
        type: string
      - description: Description of the product
        in: formData
        name: description
        required: true
        type: string
      - description: price of the product
        in: formData
        name: price
        required: true
        type: number
      - description: quantity of the product
        in: formData
        name: quantity
        required: true
        type: integer
      - description: category id of the product
        in: formData
        name: category_id
        required: true
        type: integer
      - description: Image file of the product
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ProductFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - products
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: register user
        in: body
        name: dtos.RegisterUserInput
        required: true
        schema:
          $ref: '#/definitions/dtos.RegisterUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.UserRegisterFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Register a new user
      tags:
      - users
  /api/v1/users/fetch:
    get:
      description: Fetch user login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Fetch user login
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"